update doxygen to new module dirs
