+++ doc/cctools.doxygen.config
@@ -903,11 +903,10 @@ INPUT                  = cctools.doxygen \
                          ../batch_job/src/ \
                          ../taskvine/src/manager/taskvine.h \
                          ../taskvine/src/manager/taskvine_json.h \
-                         ../taskvine/src/bindings/python3/cctools/taskvine/factory.py \
-                         ../taskvine/src/bindings/python3/cctools/taskvine/file.py \
-                         ../taskvine/src/bindings/python3/cctools/taskvine/manager.py \
-                         ../taskvine/src/bindings/python3/cctools/taskvine/task.py \
-                         ../taskvine/src/bindings/python3/cctools/taskvine/utils.py \
+                         ../taskvine/src/bindings/python3/ndcctools/taskvine/__init__.py \
+                         ../taskvine/src/bindings/python3/ndcctools/taskvine/manager.py \
+                         ../taskvine/src/bindings/python3/ndcctools/taskvine/task.py \
+                         ../taskvine/src/bindings/python3/ndcctools/taskvine/dask_executor.py \
                          ../work_queue/src/work_queue.h \
                          ../work_queue/src/bindings/python3/work_queue.binding.py \
                          ../chirp/src/chirp_reli.h \
@@ -1129,13 +1128,13 @@ STRIP_CODE_COMMENTS    = YES
 # entity all documented functions referencing it will be listed.
 # The default value is: NO.
 
-REFERENCED_BY_RELATION = YES
+REFERENCED_BY_RELATION = NO
 
 # If the REFERENCES_RELATION tag is set to YES then for each documented function
 # all documented entities called/used by that function will be listed.
 # The default value is: NO.
 
-REFERENCES_RELATION    = YES
+REFERENCES_RELATION    = NO
 
 # If the REFERENCES_LINK_SOURCE tag is set to YES and SOURCE_BROWSER tag is set
 # to YES then the hyperlinks from functions in REFERENCES_RELATION and
