+++ taskvine/src/bindings/python3/ndcctools/taskvine/__init__.py
@@ -3,31 +3,48 @@
 # See the file COPYING for details.
 
 ##
-# @package taskvine
+# @namespace ndcctools.taskvine
 #
 # Python API for the TaskVine workflow framework.
 #
 # TaskVine is a framework for building large scale distributed data intensive
 # applications that run on clusters, clouds, grids, and similar distributed systems.
-# A TaskVine application consists of a main program that creates a @ref Manager object,
-# and then submits @ref Task objects that use @ref File objects representing data sources.
+# A TaskVine application consists of a main program that creates a Manager object,
+# and then submits Task objects that use File objects representing data sources.
 # The manager distributes tasks across available workers and returns results to
 # the main application.
 #
+#
 # See the <a href=http://cctools.readthedocs.io/en/latest/taskvine>TaskVine Manual</a> for complete documentation.
 #
-# - @ref Manager
-# - @ref Task / @ref PythonTask / @ref FunctionCall
-# - @ref File
-# - @ref Factory
+# Recommended import statement:
 #
-# The objects and methods provided by this package correspond closely
-# to the native C API.
+# @code
+# import ndcctools.taskvine as vine
+# @endcode
+#
+# Relevant classes using recommended import statement:
+#
+# - @ref ndcctools.taskvine.manager.Manager "vine.Manager"
+# - @ref ndcctools.taskvine.task.Task "vine.Task"
+# - @ref ndcctools.taskvine.task.PythonTask "vine.PythonTask"
+#
+# Severless execution:
+#
+# - @ref ndcctools.taskvine.task.LibraryTask "vine.LibraryTask"
+# - @ref ndcctools.taskvine.task.FunctionCall "vine.FunctionCall"
+#
+# Dask execution:
+#
+# - @ref ndcctools.taskvine.dask_executor.DaskVine "vine.DaskVine"
+#
 
 
-from .manager import Manager
-from .file import File
-from .factory import Factory
+from .manager import (
+    Manager,
+    Factory,
+    File,
+)
 from .task import (
     Task,
     PythonTask,
@@ -35,7 +52,6 @@ from .task import (
     LibraryTask,
     FunctionCall,
 )
-from .utils import get_c_constant
 
 
 __all__ = [
@@ -47,5 +63,4 @@ __all__ = [
     "Task",
     "PythonTask",
     "PythonTaskNoResult",
-    "get_c_constant"
 ]
