--- taskvine/src/bindings/python3/ndcctools/taskvine/scheds/__init__.py
@@ -1,15 +1,0 @@
-# Copyright (C) 2023- The University of Notre Dame
-# This software is distributed under the GNU General Public License.
-# See the file COPYING for details.
-
-from .dask_executor import DaskVine
-from .dag import (
-    Dag,
-    DagNoResult
-)
-
-__all__ = [
-    "DaskVine",
-    "Dag",
-    "DagNoResult",
-]
